perf: ini user op
